more RAM targeted refactor, faster shutdown
