+++ i18n/translations/pt-BR.json
@@ -128,7 +128,7 @@
     "Missing signature file": "Arquivo de assinatura faltando",
     "Mnemonic": "Mnemônico",
     "Mnemonic ID": "ID do mnemônico",
-    "Mnemonic Storage ID": "ID de armazenamento mnemônico",
+    "Mnemonic Storage ID": "ID de armazenamento",
     "Mnemonic was not decrypted": "Mnemônico não foi descriptografado",
     "Mnemonic was not encrypted": "Mnemônico não foi criptografado",
     "Modified: ": "Modificado:",
