+++ src/krux/pages/settings_page.py
@@ -60,6 +60,7 @@ CATEGORY_SETTING_COLOR_DICT = {
 
 class SettingsPage(Page):
     """Class to mange settings interface"""
+
     def __init__(self, ctx):
         super().__init__(ctx, None)
         self.ctx = ctx
